pyproject.toml
[tool.poetry]
name = "sms_alerts"
version = "0.1.0"
description = ""
authors = ["Terje Espedal <terje.m.espedal@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
twilio = "^6.52.0"
python-dotenv = "^0.15.0"
APScheduler = "^3.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
